+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 1.1.0 - 2016-01-10
+- Moved the CLI to "packages\Nca.Valdr.1.1.0\tools\Nca.Valdr.Console.exe"
+- Created an API under Nca.Valdr.dll for the metadata parser
+
 ## 1.0.0 - 2016-01-02
 - Initial release